* commit '35a42b49e6514aeb3a164f8ab84b995f18faeeb4':
  cast to int

Conflicts:
	src/font_engine_freetype.cpp

+++ src/text/renderer.cpp
@@ -258,7 +258,7 @@ void grid_text_renderer<T>::render(glyph_positions_ptr pos, value_integer featur
                            feature_id,
                            bit->left,
                            height - bit->top,
-                           halo_radius);
+                           static_cast<int>(halo_radius));
         }
     }
 }
